refactor: i18n and view labels
Move i18n code into separate file.
Add labels for views.
Use route meta information is AppBar.vue.
