+++ src/components/AppBar.vue
@@ -120,7 +120,7 @@ export default class Layout extends AvatarMixin {
   }
 
   get currentRoute(): string | null | undefined {
-    return this.$route.name;
+    return this.$route.meta.label;
   }
 }
 </script>
